examples/fastapi-reference-project/library-api-with-oauth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.43
pydantic
pydantic-settings
python-jose[encryption]==3.3.0
httpx==0.25.2
python-dotenv==1.0.0
async-lru==2.0.4